Add multiple caption support
Make it possible to save all the captions together with the media in the
database. When serving captions the default one is being sent (has same name
as media file by default), or the next available one.
The idea/vision is to have a webinterface where users can select which captions
to use akin to what Plex has. Unfortunately it seems that UPNP clients do not support
multiple external caption files. Really a shame.