Add support for multiple album art sizes
Save album art/thumb as found in the source and return capability of sending JPEG_SM, JPEG_MED additionally to JPEG_TN in the SOAP response. The webservice will create and cache the new thumbnail on the fly if not available.
Note: as JPEG_TN is most likely to be used, always pre-generate those thumbs.