Add (periodic) rescan capability
Add a 'rescan_option' setting to the configuration file that sets the interval in minutes to start a media library rescan. Useful if inotify is not available or the library is on a path whose changes inotify can't process (e.g. NFS, FUSE).

This change also consolidated a lot of the code from inotify so the same logic is used everywhere. The way we find new/changed files is to check every path recursively from media_dirs and see whether they exist in the DB and what their timestamp is. To check for removed files we need to do another loop, this time through the database to verify everything there is still accessible.