scanner: keep timestamps of various metadata (albumart, caption, playlist, nfo)
We need to keep track of the timestamps of various metadata files. If we do not, periodic rescans will keep updating the same files over and over again. Playlists are a bit borked as the old file is removed first before it is "overwritten".

Update database version.

In the process fix an issue with captions that would cause sql errors when they are changed. Also refactor a bit how the scandir filter methods work; much more readable